Removing the unused for years app shortcuts feature.
The shortcuts feature has been commented out since 2011. Now the setings
provider no longer has a bookmarks table since the only party aside of this
commented out feature using it was the ShortcutManager. We moved to an
in memory model for the settings provider's state. We do not want to keep
there anything that is not used. Therefore, we moved from a bookmarks table
to a resource which renders this settings feature unusable. If we decide
to revive the shortcut feature we can revert this change and add support
for in memory bookmarks state in the settings provider.

Change-Id: Ib1871f940bc27029a5dd941b901d6087914a2231