Raise priority of key settings actions.
There are a bunch of actions settings handles that are really
part of the platform API, allowing apps to launch various parts
of system UI to have the user do things.  Currently these can
be intercepted by app developers by implementing the same
intent actions -- which will still have the user end up in
the activity picker to select which of the options to use, but
we really shouldn't be allowing them to get in the middle of
this at all.

To fix this, all of the specific actions that are handled by
settings and shouldn't be intercepted now have their priority
raised to 1, so they always override any filters from third
party applications.

Change-Id: I0f7050212beaec013e5ec08899d49f120ce5c249