Hide or show APNs with type="ims" as CarrierConfig
APN with type="ims" is used to attach IMS APN to the modem.
This APN is meaningless and confusing to end users.
Most carriers don't care to show this APN from apn list.
Some carriers want to show this APN from apn list and others don't.
This change hides or shows this as CarrierConfig from the APN list.

Bug: 23770219
Change-Id: Ia632f39401d84e8c63ac96f1c3b12aa3d8d2a4e6