Modify "Language & Input" for two steps selection: language, then region
The picker in frameworks/base does the simpler job: when called it
returns one locale only, no knowledge of locale lists.

The previous com.android.internal.app.LocalePicker shows a list of
locales based on the localised resources available.

The replacement (LocalePickerWithRegion) shows a more complete list
(a curated subset of all the ICU locales, but still pretty big).

Because of the length of the list the LocalePickerWithRegion splits the
selection in two: you first select the language, and then the region
(where that language is used).

If you need to select one of the locales completely supported by the
system (that have translated UI), continue to use
com.android.internal.app.LocalePicker

They should be compatible: they show some UI and end up selecting a
locale.
Beyond the UI differences, LocalePickerWithRegion will return more
"granular" locales (tens of French, or Spanish, or English locale
 variants).

Bug: 25800339
Change-Id: I0371d82302b462b1a633c149df08a39b6d508f58