Prompt user to enable chosen WebView provider if disabled.
The WebView loading logic will not allow for the user to pick a provider
that is disabled. But in the WebView developer setting for choosing
provider disabled WebView providers are shown. Therefore if the user
wants to use a disabled provider we (with this patch) prompt them to
enable the provider (if they agree we then enable it).

Bug: 26400585

Change-Id: I42476acc54f01cf873b267dd22079c5c50fa07a2