Adding an easy uninstall for active device admin
Currently the uninstall button is disabled for a package
with an active device admin. This change enables the button,
which when clicked gives the user an option to deactivate
all the DAs in the package and then uninstall the package.

Bug: b/22359208

cherry pick of I8b955305927751185a4c982dadb5b1b6b07efe5e

Change-Id: Ib2ab0e35b74fc8dba7168174b2bc8fd383fe94fe