Make Work profile lock unification a toggle
The setting to unify work and personal locks is now a toggle,
and therefore always available.

When unifying, if the work lock is sufficient as a device lock
we confirm both locks and set the work lock as the device lock.
If the work lock is not sufficient for the device, the user is
informed and taken to the lock setup flow to set a new unified
lock for both.

Also fix related Strings given UX feedback.

Bug: 26816580
Change-Id: Id019234b0a5ff4462c5f9ba86e53e30981a03e0d