Don't show disabled packages in the WebView implementation Dev setting.
You can't choose to use a disabled WebView package as WebView
implementation (this ensures that without interference from the user the
system will always try choose a package which will eventually become
updated, rather than a package that does not receive updates) and there
are cases where a package cannot/should not be enabled (to save e.g.
bandwidth).
Therefore, with this change we no longer show disabled WebView packages
as potential WebView implementations in the WebView implementation
Developer setting.

Bug: 27188851

Change-Id: I19a9b16118703d8a54d8215c186fc99ffefc4b6d