Block the user from enabling or disabling webview fallback packages.
Now that we support webview fallback packages - packages that should be
enabled if and only if no other webview packages are available - we need
to ensure that the Settings UI consistently shows that these packages
cannot be enabled or disabled (e.g. the 'Enable' and 'Disable' buttons
for enabling/disabling them are greyed out).

Also, remove the Dialog that lets a user enable a disabled webview
package from the webview implementation Dev Setting. Instead show a
Toast if the user has chosen an invalid package.

Bug: 26375524, 26375860

Change-Id: I949083d3f7c83cd2e049dd2c5c15ec5ab880fe07