Use CheckableLinearLayout for Use USB for dialog's item.
The root of "Use USB for" dialog's item is CheckableLinearLayout, which
defers the check state to direct children. However previously we had
CheckedTextView not just under CheckableLinearLayout. The structure
looks like:

Root CheckableLinearLayout -> LinearLayout -> CheckedTextView

The CL replaces the second child LinearLayout with CheckableLinearLayout
so that we can defer the state to CheckedTextView.

BUG=27708922

Change-Id: If0491235eac0a82bcb07e31856aa6b9535930fb2