Disable WebViewSetting for non-admin users.
For an app to be able to know whether the WebViewSetting is enabled for
a user we disable the setting at boot for non-admin users. In this way
an app can call intent.resolveActivity[Info] for the WebViewSetting
intent to check whether the WebViewSetting is enabled before linking to
it.

Bug: 28034166
Change-Id: I33b3fa10a38f2a2600cddd0891ef746126abdd61