Fix Work Challenge / personal confirm credential is shown on top of keyguard after rotation
ConfirmDeviceCredentialBaseFragment.SHOW_WHEN_LOCKED has been abused.
We should not set onfirmDeviceCredentialBaseFragment.SHOW_WHEN_LOCKED
for work challenge and most other cases.

It's temp fix for the rotation bug. Because it's inconsistent the activity is shown
on top of keyguard with/without rotation (or any config change). In long term,
SHOW_WHEN_LOCKED should be set only if it's truly on purpose, and remove
isKeyguardLocked checking.

Bug: 28878075
Change-Id: I19db913f5bcaf2644c8c46c25e9584103d7fee8a